+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.61 (Pending):
+## 0.61:
 * Support setRollbackOnly(), getRollbackOnly() and getTransactionStatus() on TestTransactionSynchronizationRegistry.
 * Move to using embedded payara rather than embedded glassfish.
 * Expose port in test MQ instance via OpenMQContainer.getPort().